requirements.txt
python-dotenv>=1.0.0
pyodbc>=5.0.0
Faker>=25.0.0
findspark>=2.0.0
pyspark>=3.5.0
delta-spark>=3.0.0
